IEnumerable: Use yield for large size results
Teach Permutation(), Combination(), and Scrabble() methods to work
for results with large size by using yield.

With items cached,

    var atoz = "abcdefghijklmnopqrstuvwxyz"
    atoz.Combine(25)    // returns 26 items

would take ~2s, but takes 15s-40s with iterator.

But with iterator,

    atoz.Permute(8)     // returns 62,990,928,000 items

is possible and not even possible with items cached.

For Permututation, iterator requires O(r) space, and items cached
nPr space.

So err on the side of iterator.

+++ rm.ExtensionsTest/IntExtensionTest.cs
@@ -14,7 +14,7 @@ namespace rm.ExtensionsTest
 		[TestCase(12, 479001600)]
 		public void Factorial01(int n, int result)
 		{
-			Assert.AreEqual(result, n.Factorial());
+			Assert.AreEqual(result, (int)n.Factorial());
 		}
 
 		[Test]
@@ -37,7 +37,7 @@ namespace rm.ExtensionsTest
 		[TestCase(0, 0, 1)]
 		public void Permutation01(int n, int r, int result)
 		{
-			Assert.AreEqual(result, n.Permutation(r));
+			Assert.AreEqual(result, (int)n.Permutation(r));
 		}
 
 		[Test]
@@ -46,7 +46,7 @@ namespace rm.ExtensionsTest
 		[TestCase(0, 0, 1)]
 		public void Combination01(int n, int r, int result)
 		{
-			Assert.AreEqual(result, n.Combination(r));
+			Assert.AreEqual(result, (int)n.Combination(r));
 		}
 
 		[Test]
@@ -55,7 +55,7 @@ namespace rm.ExtensionsTest
 		[TestCase(0, 0)]
 		public void ScrabbleCount01(int n, int result)
 		{
-			Assert.AreEqual(result, n.ScrabbleCount());
+			Assert.AreEqual(result, (int)n.ScrabbleCount());
 		}
 
 		[Test]
